refactor!: tabs component supports scrollable behavior
