+++ README.md
@@ -18,9 +18,8 @@ Exadel Smart Library (**ESL**) is a web components based library that gives you 
 - ##### [ESL Base Trigger](src/modules/esl-base-trigger/README.md)
 - ##### [ESL Popup](src/modules/esl-popup/README.md)
 - ##### [ESL Trigger](src/modules/esl-trigger/README.md)
-- ##### [ESL Panel and Panel Stack](src/modules/esl-panel/README.md)
-- ##### [ESL Tab and Tab Container](src/modules/esl-tab/README.md)
-- ##### [ESL Scrollable Tabs](src/modules/esl-scrollable-tabs/README.md)
+- ##### [ESL Panel and Panel Group](src/modules/esl-panel/README.md)
+- ##### [ESL Tab and Tabs](src/modules/esl-tab/README.md)
 - ##### [ESL Alert](src/modules/esl-alert/README.md)
 
 ### Form Components
