+++ src/modules/all.less
@@ -13,7 +13,6 @@
 @import "./esl-panel/core.less";
 
 @import "./esl-tab/core.less";
-@import "./esl-scrollable-tabs/core.less";
 
 @import "./esl-alert/core.less";
