--- src/modules/esl-scrollable-tabs/README.md
@@ -1,10 +1,0 @@
-# [ESL](../../../README.md) Scrollable Tabs
-
-Version: *1.0.0-beta*.  
-***Important Notice: the component is under beta version, it tested and ready to use but be aware of its potential critical API changes.***
-
-Authors: *Julia Murashko*.
-
-ESLScrollableTabs - is a custom element, container to wrap ESLTab instances and control them visually in a scrollable area.
-
-Description TBD.
