--- src/modules/esl-scrollable-tabs/core.less
@@ -1,2 +1,0 @@
-@import "./core/esl-scrollable-tabs.less";
-@import "./core/esl-scrollable-tabs-arrows.less";
