--- src/modules/esl-scrollable-tabs/core.ts
@@ -1,1 +1,0 @@
-export * from './core/esl-scrollable-tabs';
