+++ src/modules/esl-tab/core.less
@@ -1,2 +1,4 @@
 @import "./core/esl-tab.less";
 @import "./core/esl-tabs.less";
+@import "./core/esl-tabs.scrollable.less";
+@import "./core/esl-tabs.arrows.less";
