+++ src/modules/esl-tab/core/esl-tabs.arrows.less
@@ -1,4 +1,4 @@
-esl-scrollable-tabs {
+esl-tabs[scrollable] {
   &[has-scroll] {
     .arrow-prev,
     .arrow-next {
