+++ src/modules/esl-tab/core/esl-tabs.less
@@ -1,5 +1,5 @@
 esl-tabs {
-  .esl-tab-list {
+  .esl-tab-container {
     display: flex;
     flex-direction: row;
     flex-wrap: nowrap;
