+++ src/modules/esl-tab/core/esl-tabs.scrollable.less
@@ -1,7 +1,8 @@
-esl-scrollable-tabs {
+esl-tabs[scrollable] {
   display: flex;
   align-items: center;
-  .esl-tab-list {
+
+  .esl-tab-container {
     flex: 1 1 auto;
     overflow: hidden;
     white-space: nowrap;
@@ -10,6 +11,7 @@ esl-scrollable-tabs {
     flex-wrap: nowrap;
     position: relative;
   }
+
   .arrow-prev, .arrow-next {
     flex: 0 0 auto;
   }
