+++ test-server/views/pages/esl-scrollable-tab.html
@@ -1,7 +1,7 @@
 ---
 layout: ../layouts/basic.html
 containerCls: container
-title: ESL Tabs <span class="badge badge-warning">Beta</span>
+title: ESL Tabs (Scrollable) <span class="badge badge-warning">Beta</span>
 ---
 
 {{##content:
@@ -22,7 +22,7 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
     }
 
     @media (max-width: 767px) {
-      .esl-scrollable-tabs {
+      .esl-tabs {
         display: none;
       }
     }
@@ -35,13 +35,14 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
   <div class="col-12" dir="rtl">
     <h2> Tabs Component </h2>
     <div class="tabs">
-      <esl-scrollable-tabs class="tabs-container">
+      <esl-tabs scrollable class="tabs-container">
+        <esl-a11y-group targets="::find(esl-tab)" click-on-active></esl-a11y-group>
 
         <div class="arrow-prev" data-tab-direction="left">
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
@@ -104,10 +105,10 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-next"></span>
         </div>
 
-      </esl-scrollable-tabs>
+      </esl-tabs>
 
       <esl-panel-group accordion-transformation="@XS">
-        <esl-a11y-group targets="::parent::child(esl-trigger)"></esl-a11y-group>
+        <esl-a11y-group targets="::child(esl-trigger)"></esl-a11y-group>
 
         <esl-trigger class="card-header accordion-trigger"
                      target="::next"
@@ -262,13 +263,14 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
   <div class="col-12">
     <h2> Tabs Component </h2>
     <div class="tabs tabs-2">
-      <esl-scrollable-tabs class="tabs-container">
+      <esl-tabs scrollable class="tabs-container">
+        <esl-a11y-group targets="::find(esl-tab)" click-on-active></esl-a11y-group>
 
         <div class="arrow-prev" data-tab-direction="left">
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
@@ -331,10 +333,10 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-next"></span>
         </div>
 
-      </esl-scrollable-tabs>
+      </esl-tabs>
 
       <esl-panel-group>
-        <esl-a11y-group targets="::parent::child(esl-trigger)"></esl-a11y-group>
+        <esl-a11y-group targets="::child(esl-trigger)"></esl-a11y-group>
 
         <esl-trigger class="card-header accordion-trigger"
                      target="::next"
@@ -496,13 +498,14 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
   <div class="col-12">
     <h2> Tabs Component </h2>
     <div class="tabs tabs-3">
-      <esl-scrollable-tabs class="tabs-container">
+      <esl-tabs scrollable class="tabs-container">
+        <esl-a11y-group targets="::find(esl-tab)" click-on-active></esl-a11y-group>
 
         <div class="arrow-prev" data-tab-direction="left">
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
@@ -533,10 +536,10 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-next"></span>
         </div>
 
-      </esl-scrollable-tabs>
+      </esl-tabs>
 
       <esl-panel-group>
-        <esl-a11y-group targets="::parent::child(esl-trigger)"></esl-a11y-group>
+        <esl-a11y-group targets="::child(esl-trigger)"></esl-a11y-group>
 
         <esl-trigger class="card-header accordion-trigger"
                      target="::next"
@@ -614,13 +617,14 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
   <div class="col-12">
     <h2> Tabs Component </h2>
     <div class="tabs tabs-4">
-      <esl-scrollable-tabs class="tabs-container">
+      <esl-tabs scrollable class="tabs-container">
+        <esl-a11y-group targets="::find(esl-tab)" click-on-active></esl-a11y-group>
 
         <div class="arrow-prev" data-tab-direction="left">
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
@@ -659,10 +663,10 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-next"></span>
         </div>
 
-      </esl-scrollable-tabs>
+      </esl-tabs>
 
       <esl-panel-group>
-        <esl-a11y-group targets="::parent::child(esl-trigger)"></esl-a11y-group>
+        <esl-a11y-group targets="::child(esl-trigger)"></esl-a11y-group>
 
         <esl-trigger class="card-header accordion-trigger"
                      target="::next"
