+++ test-server/views/pages/esl-tab.html
@@ -32,7 +32,7 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
@@ -125,7 +125,7 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
                 <span class="icon-arrow-prev"></span>
               </div>
 
-              <ul class="nav nav-pills esl-tab-list" role="tablist">
+              <ul class="nav nav-pills esl-tab-container" role="tablist">
                 <li class="nav-item">
                   <esl-tab class="esl-tab nav-link"
                            active-class="active"
@@ -554,7 +554,7 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
@@ -801,7 +801,7 @@ title: ESL Tabs <span class="badge badge-warning">Beta</span>
           <span class="icon-arrow-prev"></span>
         </div>
 
-        <ul class="nav nav-pills esl-tab-list" role="tablist">
+        <ul class="nav nav-pills esl-tab-container" role="tablist">
           <li class="nav-item">
             <esl-tab class="esl-tab nav-link"
                      active-class="active"
